font: allow user specified italic, bold, italic bold fonts
Sometimes FcFontMatch does not return the font the user considers to be
the matching italic, bold or italic bold version of the one specified in
config.h. To deal with this, allow the user to specify the exact font
for italic, bold and italic bold versions. If any of these are not
provided, the previous mechanism will be used to heuristically pattern
match config.h user provided font to find the missing version.

The call to FcFontMatch isn't being avoided. It still happens. However,
now the user has the ability to specify the exact font that's expected.

Signed-off-by: Plato Kiorpelidis <kioplato@gmail.com>